runtime: have pid_t of type int, follow pthreads4w

+++ parity.runtime/sys/types.h
@@ -34,7 +34,7 @@
 #  include UCRT_INC(sys/Types.h)
 #pragma pop_macro("__STDC__")
 
-typedef long pid_t;
+typedef int pid_t;
 typedef int mode_t;
 
 #if defined(_WIN64)
